feat: disabled buttons in un protected components
