+++ src/components/ProtectedRoute/index.js
@@ -1,25 +1,16 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
 import { Link, Route } from 'react-router-dom';
 import { Result, Button } from 'antd';
+import getUserPermission from '../../utils/getUserPermission';
 
 function ProtectedRoute({ component: Component, permission: permission, ...rest }) {
-  const { resource, action } = permission;
-  const userPermission = useSelector(({ spaces }) => {
-    const { selected, details } = spaces;
-    return details[selected] ? details[selected].permissions : [];
-  });
+  const actions = getUserPermission(permission);
 
-  const node = userPermission.findIndex(
-    (each) =>
-      each.resource === 'admin' || (each.resource === resource && each.actions.includes(action)),
-  );
-
-  if (node > -1)
+  if (actions.length > 0)
     return (
       <Route
         {...rest}
-        render={(props) => <Component {...rest} {...props} permission={userPermission[node]} />}
+        render={(props) => <Component {...rest} {...props} permission={{ actions }} />}
       />
     );
